# apps/api/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Express Starter API
  description: |
    Express Starter API 文档
    
    本API提供以下功能：
    - 健康检查
    - 用户鉴权和API Key管理
    - 持久化APIKEY管理
    - 第三方API代理服务
    
    ## 鉴权方式
    
    本API支持两种鉴权方式：
    
    1. **Bearer Token（临时Token）**
       - 通过 `/api/v1/auth/api-key` 或 `/auth/api-key` 接口获取
       - 存储在内存中，服务重启后失效
       - 使用方式：`Authorization: Bearer {access_key}`
    
    2. **APIKEY（持久化Key）**
       - 通过 `/api/v1/api-keys` 或 `/api-keys` 接口创建
       - 存储在数据库中，服务重启后仍然有效
       - 支持过期时间设置和启用/禁用
       - 使用方式：
         - 请求头：`X-API-Key: {api_key}`
         - 查询参数：`?api_key={api_key}`
    
    鉴权优先级：优先使用Bearer Token，如果没有或无效，则使用APIKEY。
    
    **注意**：所有接口都支持两种路径格式：
    - 完整路径：`/api/v1/{path}` （推荐）
    - 便捷路径：`/{path}` （兼容旧版本）
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000/api/v1
    description: 本地开发环境
  - url: https://api.example.com/api/v1
    description: 生产环境
...
paths:
...
  /api-keys:
    post:
      tags:
        - ApiKeys
      summary: 创建持久化APIKEY
      description: |
        创建持久化的APIKEY，存储在数据库中，服务重启后仍然有效。
        需要先通过 `/api/v1/auth/api-key` 或 `/auth/api-key` 接口获取临时Token，然后使用Bearer Token进行鉴权。
        
        **重要**：APIKEY创建后只会在响应中显示一次，请妥善保管。
        
        **路径说明**：
        - 完整路径：`POST /api/v1/api-keys`
        - 便捷路径：`POST /api-keys`
      operationId: createApiKey
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: APIKEY名称（可选）
                  example: My Production API Key
                expires_at:
                  type: string
                  format: date-time
                  description: 过期时间（可选，ISO 8601格式）
                  example: "2026-12-31T23:59:59Z"
            examples:
              basic:
                summary: 基本请求（使用默认名称）
                value: {}
              withName:
                summary: 指定名称
                value:
                  name: Production API Key
              withExpiry:
                summary: 指定过期时间
                value:
                  name: Temporary API Key
                  expires_at: "2026-12-31T23:59:59Z"
      responses:
        '201':
          description: APIKEY创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        description: APIKEY ID
                      api_key:
                        type: string
                        description: 生成的APIKEY（只显示一次）
                        example: sk_1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
                      name:
                        type: string
                        description: APIKEY名称
                      created_at:
                        type: string
                        format: date-time
                        description: 创建时间
                      expires_at:
                        type: string
                        format: date-time
                        nullable: true
                        description: 过期时间
                      message:
                        type: string
                        example: APIKEY 已生成，请妥善保管，此密钥只会显示一次
              example:
                success: true
                data:
                  id: "123e4567-e89b-12d3-a456-426614174000"
                  api_key: "sk_1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
                  name: "My Production API Key"
                  created_at: "2026-01-17T12:00:00Z"
                  expires_at: null
                  message: "APIKEY 已生成，请妥善保管，此密钥只会显示一次"
        '401':
          description: 鉴权失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    get:
      tags:
        - ApiKeys
      summary: 获取当前用户的APIKEY列表
      description: |
        获取当前已登录用户的所有APIKEY列表。
        APIKEY会部分隐藏，只显示前8位和后4位。
      operationId: listApiKeys
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          description: APIKEY ID
                        api_key:
                          type: string
                          description: 部分隐藏的APIKEY
                          example: sk_1234...abcd
                        name:
                          type: string
                          description: APIKEY名称
                        is_active:
                          type: boolean
                          description: 是否激活
                        last_used_at:
                          type: string
                          format: date-time
                          nullable: true
                          description: 最后使用时间
                        expires_at:
                          type: string
                          format: date-time
                          nullable: true
                          description: 过期时间
                        created_at:
                          type: string
                          format: date-time
                          description: 创建时间
              example:
                success: true
                data:
                  - id: "123e4567-e89b-12d3-a456-426614174000"
                    api_key: "sk_1234...abcd"
                    name: "My Production API Key"
                    is_active: true
                    last_used_at: "2026-01-17T12:30:00Z"
                    expires_at: null
                    created_at: "2026-01-17T12:00:00Z"
        '401':
          description: 鉴权失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api-keys/{id}:
    delete:
      tags:
        - ApiKeys
      summary: 删除APIKEY
      description: 删除指定的APIKEY，删除后该APIKEY将无法再使用。
      operationId: deleteApiKey
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: APIKEY ID
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "APIKEY 已删除"
              example:
                success: true
                message: "APIKEY 已删除"
        '401':
          description: 鉴权失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: APIKEY不存在或无权删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api-keys/{id}/toggle:
    patch:
      tags:
        - ApiKeys
      summary: 启用/禁用APIKEY
      description: 启用或禁用指定的APIKEY，禁用后的APIKEY将无法用于鉴权。
      operationId: toggleApiKey
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: APIKEY ID
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - is_active
              properties:
                is_active:
                  type: boolean
                  description: 是否激活
                  example: false
            examples:
              disable:
                summary: 禁用APIKEY
                value:
                  is_active: false
              enable:
                summary: 启用APIKEY
                value:
                  is_active: true
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "APIKEY 已禁用"
              example:
                success: true
                message: "APIKEY 已禁用"
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  value:
                    success: false
                    message: "is_active 必须是布尔值"
                    error: {}
        '401':
          description: 鉴权失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: APIKEY不存在或无权操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        临时Bearer Token鉴权方式。
        通过 `/auth/api-key` 接口获取的access_key作为Bearer token。
        格式：`Authorization: Bearer {access_key}`
        注意：此Token存储在内存中，服务重启后会失效。
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        持久化APIKEY鉴权方式。
        通过 `/api-keys` 接口创建的持久化APIKEY。
        支持两种方式：
        1. 请求头：`X-API-Key: {api_key}`
        2. 查询参数：`?api_key={api_key}`
        注意：APIKEY存储在数据库中，服务重启后仍然有效。
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: 错误消息
        error:
          type: object
          description: 错误详情
security:
  - BearerAuth: []